cast times to unsigned long before printing with %lu.
I was getting warnings about integer type, long format arg, and POSIX
says only that time_t is integral or floating (!!).  It seems it's
always integer, long seems big enough, and the code already assumed
unsigned.

(in test code and clearly a bug, so committing even though we're in a
semi-freeze)

+++ test_mkgmtime.c
@@ -97,7 +97,7 @@ int main(int argc, char *argv[])
 			(void)strftime(tbuf, sizeof(tbuf), "%F %T", &tests[i].t);
 			(void)printf("test %2d failed. "		\
 			       "Time returned from: %s should be %lu (but was: %lu)\n",
-				i, tbuf, tests[i].result, ts);
+				i, tbuf, (unsigned long) tests[i].result, (unsigned long) ts);
 		}
 	}
 	return (int)failed;
